+++ src/components/drawer/HomeDrawer.vue
@@ -6,18 +6,31 @@
     data-cy="home-drawer"
   >
     <template #content>
-      <q-btn
-        icon="fa-solid fa-clock-rotate-left"
-        flat
-        color="accent"
-        no-caps
-        :label="$t('page.home.project.recent')"
-        align="left"
-      />
+      <q-list>
+        <q-item
+          v-ripple
+          clickable
+          active
+          active-class="text-accent"
+        >
+          <q-item-section avatar>
+            <q-icon name="fa-solid fa-clock-rotate-left" />
+          </q-item-section>
+          <q-item-section class="text-bold">
+            {{ $t('page.home.drawer.projects.recent') }}
+          </q-item-section>
+        </q-item>
+        <short-projects-item />
+        <q-separator
+          color="primary"
+          class="q-ma-md"
+        />
+      </q-list>
     </template>
   </default-drawer>
 </template>
 
 <script setup>
 import DefaultDrawer from 'src/components/drawer/DefaultDrawer';
+import ShortProjectsItem from 'src/components/item/ShortProjectsItem';
 </script>